Add validation function for DSSE, separate c2pa manifest signer
